% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoGloBias.R
\name{calculate.statistics}
\alias{calculate.statistics}
\title{A function to group observations and calculate statistics}
\usage{
calculate.statistics(
  ldf,
  group.by = NULL,
  collapse = NULL,
  statistics = list(mean = "mean", med = "median", SD = "sd", IQR = "IQR", MAD = "mad",
    skewness = "skewness")
)
}
\arguments{
\item{ldf}{Long format dataframe with one column named 'value' containing the values, one identifier-column specifying type of the value (such as a column defining the metric), and at least one identifier-column that can be used for grouping (such as participant id).}

\item{group.by}{Name of the identifier-columns that are used to to group the observations. Statistics are calculate for each grouping separately, e.g. group.by = c('sessions', 'measures') means that the statistic will be calculated separately for each session and each measure. Is mutually exclusive with collapse. Defaults to NULL.}

\item{collapse}{Name of the identifier-columns that are NOT used to to group the observations. E.g. collapse = c('sessions','participant id') together with correlate.by = 'metric' means that the metrics will be correlated for each session separately, but rather be pooled and 'session' vanishes as independent grouping in the results. Is mutually exclusive with group.by. Defaults to NULL.}

\item{statistics}{List of statistics to be applied to the groupings. The name of a list element is the statistic-name and the element defines the name of the function used to calculate it. Defaults to 'list(mean = 'mean', med = 'median', SD = 'sd', IQR = 'iqr', MAD = 'mad', skewness = 'skewness')'.}
}
\description{
A function to group observations in a long format dataframe and calulate a range of statistics.
}
\examples{
# Summarise trials to trial types, then calculate scores
scores = score.summaries(ldf = summarise.trials(ldf = example.performance.data()))
# Calculate statistics in the score-ldf by collapsing the id-column, i.e. for all other combinations of columns (except value) the statistic will be calculated separately
calculate.statistics(ldf = scores, collapse = 'id')
# With the given data in scores, this is equivalent to grouping by session and measure
calculate.statistics(ldf = scores, group.by = c('session', 'measure' ))
}
\keyword{statistic}
\keyword{summary,}
