% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoGloBias.R
\name{correlate.values}
\alias{correlate.values}
\title{A function to correlate observations}
\usage{
correlate.values(
  ldf,
  correlate.by = "metric",
  group.by = NULL,
  collapse = NULL,
  method = "pearson",
  notify = T
)
}
\arguments{
\item{ldf}{Long format dataframe with one column named 'value' containing the values, one identifier-column specifying type of the value (such as a column defining the metric), and at least one identifier-column that can be used for grouping (such as participant id).}

\item{correlate.by}{Name of the column containing the identifiers that should be correlated with each other. Defaults to 'metric'.}

\item{group.by}{Name of the identifier-columns that are used to to group the observations. Each grouping is correlated internally by 'correlate.by'. E.g. group.by = c('sessions', 'measures') together with correlate.by = 'metric' means that metrics will be correlated with each other, for each session and each measure separately. Is mutually exclusive with collapse. Defaults to NULL.}

\item{collapse}{Name of the identifier-columns that are NOT used to to group the observations. E.g. collapse = c('sessions','participant id') together with correlate.by = 'metric' means that the metrics will be correlated for each session separately, but rather be pooled and 'session' vanishes as independent grouping in the results. Is mutually exclusive with group.by. Defaults to NULL.}

\item{method}{Method to be used by cor.test(). Choices are 'pearson', 'spearman', and 'kendall'. Defaults to 'pearson'.}

\item{notify}{Provides progress information. Defaults to TRUE.}
}
\value{
s statistic, e.g. t-statistic for pearson

n number, degrees of freedom for pearson

p significance, p-value

c correlation, e.g. r for pearson

CI1,CI2 95% confidence interval
}
\description{
A function to correlated observations in a long format dataframe, separately for chosen groupings. Returns the results of the correlation, as specified by the cor-column.
}
\examples{
# Summarise trials to trial types, then calculate scores
scores = score.summaries(ldf = summarise.trials(ldf = example.performance.data()))
# Correlate metrics in the score-ldf by collapsing the id-column, ie. for all other combinations of columns (except metric and value) the correlations will be calculated separately
correlate.values(ldf = scores, correlate.by = 'metric', collapse = 'id')
# With the given data in scores, this is equivalent to grouping by session and measure
correlate.values(ldf = scores, correlate.by = 'metric', group.by = c('session', 'measure' ))
}
\keyword{summary}
