% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoGloBias.R
\name{default.metrics}
\alias{default.metrics}
\title{Return the default metric instructions}
\usage{
default.metrics()
}
\description{
A function that just returns hard-coded calculation instructions for the default metrics.
}
\examples{
default.metrics()
}
\keyword{score}
