% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoGloBias.R
\name{example.performance.data}
\alias{example.performance.data}
\title{Create non-representative example data for a performance-focused task}
\usage{
example.performance.data(
  mean.RT = 400,
  sd.RT = 150,
  mean.accuracy = 0.9,
  level = c(rep("loc", 5), rep("glo", 5)),
  congruence = c("con", "inc"),
  id = 1:10,
  session = 1:2,
  format = "wide",
  seed = NULL
)
}
\arguments{
\item{mean.RT, sd.RT}{Overall mean and standard deviation values being passed to rnorm() to create example reaction times. Any negative values generated by rnorm() will be made positive. Default to 400 and 150, respectively.}

\item{mean.accuracy}{Overall mean accuracy used to create example accuracies. Defaults to 0.9.}

\item{level}{A vector of level identifiers, describing the level of EACH trial in an experiment. Is being used to determine (together with the length of 'congruence') the number of trials for each participant, for each session. Defaults to c(rep('loc',5), rep('glo',5)).}

\item{congruence}{A vector of level identifiers, describing which KINDS of congruence exists. Is being used to determine (together with the length of 'level') the number of trials. Defaults to c('con','inc').}

\item{id, session}{Vectors of participant and session identifiers. Default to 1:8 and 1:2 respectively.}

\item{format}{Defines if a true long format dataframe ('long') or a long-wide hybrid ('wide') is returned. Defaults to 'wide'.}
}
\description{
Create example trial-level local-global data from experiment in which a participant's performance is evaluated. The data does not represent any real patterns and merely severs to demonstrated the required structure.
}
\examples{
example.performance.data()
}
\keyword{data}
\keyword{example,}
