% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoGloBias.R
\name{example.preference.data}
\alias{example.preference.data}
\title{Create non-representative example data for a preference-focused task}
\usage{
example.preference.data(
  mean.RT = 400,
  sd.RT = 150,
  mean.preference = 0,
  congruence = c(rep("con", 5), rep("inc", 5)),
  id = 1:8,
  session = 1:2,
  format = "wide",
  seed = NULL
)
}
\arguments{
\item{mean.RT, sd.RT}{Overall mean and standard deviation values being passed to rnorm() to create example reaction times. Any negative values generated by rnorm() will be made positive. Default to 400 and 150, respectively.}

\item{mean.preference}{Overall mean preference used to create example preferences. Defaults to 0.5, which means equal chance of having a local or global preference.}

\item{congruence}{A vector of level identifiers, describing which KINDS of congruence exists. Is being used to determine (together with the length of 'level') the number of trials. Defaults to c('con','inc').}

\item{id, session}{Vectors of participant and session identifiers. Default to 1:8 and 1:2 respectively.}

\item{format}{Defines if a true long format dataframe ('long') or a long-wide hybrid ('wide') is returned. Defaults to 'wide'.}
}
\description{
Create example trial-level local-global data from experiment in which a participant's preference is evaluated. The data does not represent any real patterns and merely severs to demonstrated the required structure.
}
\examples{
example.preference.data()
}
\keyword{data}
\keyword{example,}
