% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoGloBias.R
\name{summarise.trials.PREF}
\alias{summarise.trials.PREF}
\title{Summarise trials of a local-global preference bias experiment}
\usage{
summarise.trials.PREF(
  wdf,
  group.by = NULL,
  group.by0 = NULL,
  calculate = data.frame(mean_PREF = c("PREF", "mean"), sum_PREF = c("PREF", "sum"),
    mean_RT = c("RT", "mean"), median_RT = c("RT", "median")),
  measure.cols = c("PREF", "RT"),
  level.col = "level",
  preference.col = "PREF",
  make.trial_type = NULL
)
}
\arguments{
\item{group.by, group.by0}{Identifier-columns that are used create trial-groupings, within which statistics are calculated. The column names defined by group.by are used to create a power set (all possible combinations, including NULL), while the column names specified by group.by0 are added to each element of the power set. The resulting sets with combinations of column names are used iteratively when grouping the data (see examples). Default to NULL.}

\item{calculate}{A 2*x dataframe with instructions how to calculate summary statistics within the groupings. Column names specify the name for the summary statistic, the first row specifies the measure to be summarised, and the second row the function to be used. The function must accept a vector and output a scalar. Defaults to data.frame(mean_ACC = c('ACC','mean'), ER = c('ER','mean'), mean_RT = c('RT','mean'), median_RT = c('RT','median')).}

\item{measure.cols}{Names of the columns containing the measures. Default to c('PREF','RT') for accuracy and reaction time. Values must be numeric. Accuracy must be coded as 1 or 0. The error rate is automatically calculated from accuracy.}

\item{level.col}{Name of the to-be-created column that will contain the chosen level (preference) in each trial. If the trials should be grouped by level (group.by/group.by0) for calculation of RTs, this name must be used for the level-column. Is ignored if it does not exist in group.by/group.by0. Defaults to 'level'.}

\item{preference.col}{Name of the column holding the preference-measure. This name must be used as measure-name in the 'calculate' instructions. Defaults to PREF.}

\item{make.trial_type}{Columns to be used to create a singular trial_type column. Can only specify 1 or 2 columns to be used (e.g. level and congruence). Defaults to NULL.}

\item{ldf}{Wide-long hybrid dataframe with the trial data. The columns defined by 'measure.cols' contain the values and the other columns are either irrelevant or are used as identifier-columns to index these values.}
}
\description{
Group trials of a local-global bias experiment in which a participant's preference instead of accuracy is measured. Group trials in a long format dataframe and apply defineable summary statistic functions to the measures. A trial's level is defined by the choice the participant made.
}
\examples{
# Group by the power set of session, id, level, and congruence. Use default 'calculate'-instructions.
ldf = summarise.trials.PREF(wdf = example.preference.data(), group.by = c('session','id','level','congruence'))
# Group by the power set of level and congruence and the normal set of session and id. Now, only summaries that differentiate by session and id are calculated (e.g. no overall summary over the entire data)
ldf = summarise.trials.PREF(wdf = example.preference.data(), group.by = c('level','congruence'), group.by0 = c('session','id'))
debug(summarise.trials.PREF)
summarise.trials.PREF(wdf = example.preference.data(seed = 2),group.by0 = c('session','id','level','congruence'), make.trial_type = c('level','congruence'))
}
\keyword{summary}
