% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoGloBias.R
\name{summarise.trials}
\alias{summarise.trials}
\title{Summarise trials of a local-global preference bias experiment}
\usage{
summarise.trials(
  wdf,
  group.by = NULL,
  group.by0 = NULL,
  fuse.by = NULL,
  correct.trials.only = T,
  calculate = data.frame(mean_ACC = c("ACC", "mean"), mean_ER = c("ER", "mean"), mean_RT
    = c("RT", "mean"), median_RT = c("RT", "median")),
  measure.cols = c("ACC", "RT"),
  accuracy = "ACC",
  error_rate = "ER",
  make.trial_type = NULL
)
}
\arguments{
\item{group.by, group.by0}{Identifier-columns that are used create trial-groupings, within which statistics are calculated. The column names defined by group.by are used to create a power set (all possible combinations, including NULL), while the column names specified by group.by0 are added to each element of the power set. The resulting sets with combinations of column names are used iteratively when grouping the data (see examples). Default to NULL.}

\item{correct.trials.only}{Removes all values that are not accuracy or error rate from trials with an accuracy of 1. Defaults to TRUE.}

\item{calculate}{A 2*x dataframe with instructions how to calculate summary statistics within the groupings. Column names specify the name for the summary statistic, the first row specifies the measure to be summarised, and the second row the function to be used. The function must accept a vector and output a scalar. Defaults to data.frame(mean_ACC = c('ACC','mean'), ER = c('ER','mean'), mean_RT = c('RT','mean'), median_RT = c('RT','median')).}

\item{measure.cols}{Names of the columns containing the measures. Default to c('ACC','RT') for accuracy and reaction time. Values must be numeric. Accuracy must be coded as 1 or 0. The error rate is automatically calculated from accuracy.}

\item{accuracy, error_rate}{The names of the columns that contain accuracy and error rate. If accuracy exists, but not error rate, the error rate is automatically calculated and gets the named defined by 'error_rate'. Default to ACC and ER.}

\item{make.trial_type}{Columns to be used to create a singular trial_type column. Can only specify 1 or 2 columns to be used (e.g. level and congruence). Defaults to NULL.}

\item{ldf}{Wide-long hybrid dataframe with the trial data. The columns defined by 'measure.cols' contain the values and the other columns are either irrelevant or are used as identifier-columns to index these values.}
}
\description{
Group trials of a local-global bias experiment in which a participant's preference instead of accuracy is measured. Group trials in a long format dataframe and apply defineable summary statistic functions to the measures. A trial's level is defined by the choice the participant made.
}
\examples{
# Group by the power set of session, id, level, and congruence. Use default 'calculate'-instructions.
ldf = summarise.trials(wdf = example.performance.data(), group.by = c('session','id','level','congruence'))
# Group by the power set of level and congruence and the normal set of session and id. Now, only summaries that differentiate by session and id are calculated (e.g. no overall summary over the entire data)
ldf = summarise.trials(wdf = example.performance.data(), group.by = c('level','congruence'), group.by0 = c('session','id'))
# Calculate the sd of RT for the entire dataset and no other summary. Also, use all RTs, not only the RTs from correct trials.
ldf = summarise.trials(wdf = example.performance.data(), calculate = data.frame(sd_of_all_RT = c('RT','sd')), correct.trials.only = F)
}
\keyword{summary}
